Use `prepare` instead of `postinstall` to initialize hooks (#4)
`node-git-hooks` is only used as devDependencies, but `postinstall` is 
also triggered when published package is installed via npm. Thus one should
also install `node-git-hooks` in their repository, or the package publisher
should include `node-git-hooks` as a dependencies, not devDependencies.

Changing `postinstall` to `prepare` prevent it, since it installs the hooks only
when developing the package, not when installing the package via npm.

Refer to the [npm docs](https://docs.npmjs.com/cli/v7/using-npm/scripts#prepare-and-prepublish) for more detail.